moved categories state to app to have access in post
