+++ src/Components/NavBar.jsx
@@ -2,9 +2,8 @@ import { useState, useEffect } from "react"
 import { getCategories } from "../apis";
 import { Link } from 'react-router-dom';
 
-export default function NavBar () {
+export default function NavBar ({categories, setCategories}) {
 
-    const [categories, setCategories] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
 
     useEffect(() => {
@@ -12,7 +11,7 @@ export default function NavBar () {
             setCategories(data.categories);
             setIsLoading(false);
         });
-    }, []);
+    }, [setCategories]);
 
     if (isLoading) return <p>Loading Categories...</p>
