+++ src/css/App.css
@@ -126,5 +126,4 @@ li {
     .single-review {
       max-width: 1450px;
     }
-        
 }
